Another registration modification
This commit builds on the previous by one again moving where the
resource is registered. The container is now registered with Terraform
after it's created but before it's started. This should be a good
balance between preventing erroneous containers while being able to
clean up containers that failed to launch.